Refactor database configuration in auth-plugin
This commit refactors the database configuration in the auth-plugin. The `DB_TYPE` environment variable is now set to the default value of "sqlite" if not provided. The `DB_URL` environment variable is set to the default value of "sqlite:///./test.db" if not provided. This change improves the readability and maintainability of the code.